Check URI equality a bit more loosely for menu's active check
This is initally to fix active page menus like "Edit Search" on search
pages. Previously even if the URL in browser is the same as the page
menu, the page menu wouldn't be active(shown as bold).

There are 3 different ways to genearte URI: CGI, URI and manually. URI
strings they generate could be different in character perspective but
actually identical in URI perspective, and the difference is mainly in
the query part.

E.g.

* Delimiter

CGI uses ";" and URI uses "&"

We can customize it via $URI::DEFAULT_QUERY_FORM_DELIMITER in URI, but
sadly delimiter is not the only difference.

* Space in query value

CGI uses "%20" and URI uses "+"

We are comparing difference resources(the one from CGI and the other
from URI or manual) when determining if one menu is active or not. This
commit compares them using clean URI objects, to eliminate the chaos
above.

See also 07c22ca552, which is an earlier attempt but not complete.