List current user in "All Recipients" if it's a watcher
Previously when NotifyActor is false, even if current user is a watcher,
it was not listed in "Modify who receives mail ..." widget on modify
people page, which was confusing as current user would get notifications
of actions made by others.

This commit fixes this, with a hint to indicate that RT won't send
notifications of its own actions.

This also allows People to ignore specific ticket updates by unchecking
themselves there.