Warn on invalid result formats in Asset System Configuration
The options AssetDefaultSearchResultFormat, AssetSimpleSearchFormat,
AssetSummaryFormat, AssetSummaryRelatedTicketsFormat, and
UserAssetExtraInfo are intended to contain search result format
specifications for their respective objects for display in specific
places.

This change modifies the System Configuration web UI pages to add
warnings to the update results when portions of a changed format don't
appear to map to any field or attribute of the related object type.
This doesn't necessarily mean that there is an error, but it is unusual
enough to be worth bringing to the user's attention.

It is possible to inadvertently introduce small, subtle, errors into
these strings that prevent them from showing any useful information.
While this can also happen in configuration files,

This is not applied to values set in configuration files.