Display new and old values over max lines in two divs
When displaying the new and old values for a SetConfig or CustomField text transaction, split values that have more lines than the MaxLines parameter into two divs. One div displays the value truncated to the max number of lines and one div displays the full value. The divs use the bootstrap collapse class and collapse buttons to toggle display between the two divs.