Add callback for modifying which CFs are diffed
In the ticket transaction display by default custom fields with a type
that matches /text/ are show collapsed with a widget to expand and see a
diff of the old and new contents.

This change adds a callback, /Elements/ShowTransaction/ModifyShowCFDiff,
that allows direct control over when custom fields are shown with a
collapsed diff and when they are shown normally.