Make the Timezone config option a Select widget
This change converts the Timezone option to a Select widget, from a text
box, on the System Configuration page.  This restricts the admin to
choosing from known-valid options.

This is the same list that is built and presented in the User
Preferences.

Note that while this list does not usually include deprecated timezone
names, the current value of the configuration option and the file-only
value of the option (if different from the current value) are both
added at the top of the list, even when they are deprecated.