Tweak spacing on ticket create/update with gnupg_widget
On the Ticket Create page the Subject is the last item in the "Create a
new ticket in ..." group before the option to include an article (if
present), but when $gnupg_widget is set there is a pair of checkboxes
displayed after the Subject.

The additional spacing after the Subject inputs, due to the "form-group"
class on the containing <div>, makes the checkboxes look like they're
closely related to the "Include Article" input when it is present, but
they are not.

A similar issue existed on the Ticket Update page, but in this case the
UI was inconsistent in that the encryption checkboxes were located above
the Subject textbox.  They have been moved to be below the Subject in
addition to fixing the spacing inconsistency.

This change detects when $gnupg_widget is set and adjusts the location
of the "form-group" class appropriately to make the spacing align better
with the logical grouping of the controls.