Sort SavedSearches
SavedSearches were previously returned in an unpredictable order
paralleling the order that the database returned the underlying
Attributes, probably based more on the order in which they were created
than on anything else.

This change automatically sorts the SavedSearches by Name (stored in the
Attribute's Description) after they are fetched in LimitToPrivacy.