Fix logic in Logout menu option
When WebRemoteUserAuth is enabled and WebFallbackToRTLogin is not enabled
the user would not see the Logout menu option.

When WebFallbackToRTLogin is enabled the user would always see the native
RT logout even if they were externally authed.

Changed the logic to use the native RT logout when using native RT login
and only use the LogoutURL config option if the user is actually
authorized externally and the LogoutURL config option is not set to the
native RT logout.